Implemented morton encoding using standard bit-twidling approach.

+++ include/PolyVox/PagedVolume.h
@@ -127,10 +127,10 @@ namespace PolyVox
 			VoxelType* getData(void) const;
 			uint32_t getDataSizeInBytes(void) const;
 
-			VoxelType getVoxel(uint16_t uXPos, uint16_t uYPos, uint16_t uZPos) const;
+			VoxelType getVoxel(uint32_t uXPos, uint32_t uYPos, uint32_t uZPos) const;
 			VoxelType getVoxel(const Vector3DUint16& v3dPos) const;
 
-			void setVoxel(uint16_t uXPos, uint16_t uYPos, uint16_t uZPos, VoxelType tValue);
+			void setVoxel(uint32_t uXPos, uint32_t uYPos, uint32_t uZPos, VoxelType tValue);
 			void setVoxel(const Vector3DUint16& v3dPos, VoxelType tValue);
 
 		private:
